live_demo/requirements.txt
streamlit
easyocr
opencv-python-headless
deep-translator
numpy
Pillow
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu